refactor: move getWalletBalanceForTx to commons.ts
